extract xml reporting
